[Printer][Refactoring] refactoring printer.cc
