+++ dynamic/PrinterTest.cpp
