+++ dynamic/dynamic.vcxproj.filters
@@ -21,7 +21,7 @@
     <ClInclude Include="Define.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="printer.h">
+    <ClInclude Include="Printer.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
@@ -29,10 +29,10 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="printer_test.cpp">
+    <ClCompile Include="PrinterTest.cpp">
       <Filter>소스 파일\UnitTest</Filter>
     </ClCompile>
-    <ClCompile Include="printer.cpp">
+    <ClCompile Include="Printer.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
   </ItemGroup>
